backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
transformers==4.35.2
torch==2.1.1
python-multipart==0.0.6
mysqlclient==2.2.0
python-dotenv==1.0.0 